# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asxana API",
        "description": "Asxana loyihasi uchun REST API dokumentatsiyasi. Bu API orqali siz kategoriyalar, mahsulotlar va boshqa resurslar bilan ishlashingiz mumkin.",
        "contact": {
            "name": "API Support",
            "email": "support@asxana.uz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Mahsulotlar ro'yxatini olish",
                "description": "Filtrlar yordamida mahsulotlarni saralab olish va tillar bo'yicha ma'lumotlarni ko'rish",
                "operationId": "7ca3593d05653d957ae7be1ac648ca44",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Tilni tanlash (uz, ru, kk)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Kategoriya bo'yicha filtralsh",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli yakunlandi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category_name": {
                                                        "type": "string",
                                                        "example": "Salat"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "sezar-salat.jpg"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 25000
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Sezar salat"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Tovuq go'shti, parmezan bilan klassik Sezar salat"
                                                    },
                                                    "is_available": {
                                                        "type": "boolean",
                                                        "example": true,
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-12-22T04:52:34.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-12-22T04:52:34.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve products"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Yangi mahsulot yaratish",
                "description": "Mahsulot asosiy ma'lumotlari va bir nechta tildagi tarjimalarini rasm bilan birga saqlaydi.",
                "operationId": "cfdb2a3921bee30f9cad135e1c5f1452",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "price",
                                    "image",
                                    "translations"
                                ],
                                "properties": {
                                    "category_id": {
                                        "description": "Kategoriya ID raqami",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "description": "Mahsulot narxi",
                                        "type": "integer",
                                        "example": 25000
                                    },
                                    "is_available": {
                                        "description": "Sotuvda bor yoki yo'qligi",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "image": {
                                        "description": "Mahsulot rasmi (fayl)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "translations": {
                                        "description": "Turli tillardagi nom va tavsiflar",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "lang_code": {
                                                    "description": "Til kodi (uz, ru, kk)",
                                                    "type": "string",
                                                    "example": "uz"
                                                },
                                                "name": {
                                                    "description": "Mahsulot nomi",
                                                    "type": "string",
                                                    "example": "Sezar salat"
                                                },
                                                "description": {
                                                    "description": "Mahsulot tavsifi",
                                                    "type": "string",
                                                    "example": "Tovuq go'shti va parmezan bilan"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mahsulot muvaffaqiyatli yaratildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatsiya xatosi (masalan: rasm formati noto'g'ri yoki nom band)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Serverda xatolik yuz berdi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProductResource": {
                "title": "Product Resource",
                "description": "Mahsulot ma'lumotlari modeli",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_name": {
                        "type": "string",
                        "example": "Salat"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://misol.uz/storage/products/rasm.jpg"
                    },
                    "price": {
                        "type": "integer",
                        "example": 25000
                    },
                    "name": {
                        "type": "string",
                        "example": "Sezar salat"
                    },
                    "description": {
                        "type": "string",
                        "example": "Tovuqli klassik salat"
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}